% Auto-generated: do not edit by hand
\name{''OrientationObserver}

\alias{''OrientationObserver}

\title{OrientationObserver component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
''OrientationObserver(id=NULL, before_orientation=NULL, orientation=NULL)
}

\arguments{
\item{id}{Character. }

\item{before_orientation}{Character. }

\item{orientation}{Character. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

